clip/requirements.txt
torch==2.5.1
clip @ git+https://github.com/openai/CLIP.git
ftfy==6.3.1 
regex==2024.9.11
tqdm==4.66.6
pika==1.3.2
psycopg2-binary==2.9.10
